move dictionarybuilding and quantizeEquipment functions to equipment utility function file
