+++ src/utils/equipment.ts
@@ -37,3 +37,47 @@ export function filterEquipment(
     });
   });
 }
+
+// Function to convert equipment Array to Quantized equipment Array
+export function quantizeEquipment(equipment: Equipment[]): Equipment[] {
+  const tempArray: Equipment[] = [];
+  equipment.forEach((item) => {
+    item.quantity = equipment.filter(
+      (element) => element.description === item.description
+    ).length;
+    const index = tempArray.findIndex(
+      (element) => element.description === item.description
+    );
+    if (index === -1) {
+      tempArray.push(item);
+    }
+  });
+  return tempArray;
+}
+
+// Function to convert equipment Array to Quantized equipment Array
+export function buildDictionaries(
+  equipment: Equipment[]
+): [
+  { [k: string]: boolean },
+  { [k: string]: Set<string> },
+  { [k: string]: number }
+] {
+  // Build quantities dictionary for Formik
+  // Build Categories Dictionary
+  // Build Filters Dictionary
+  const filters: { [k: string]: boolean } = {};
+  const categories: { [k: string]: Set<string> } = {};
+  const quantities: { [k: string]: number } = {};
+  equipment.forEach((item) => {
+    quantities[item.description] = 0;
+    item.tags.forEach((tag) => {
+      if (!categories[item.category]) {
+        categories[item.category] = new Set();
+      }
+      categories[item.category].add(tag.name);
+      filters[tag.name] = false;
+    });
+  });
+  return [filters, categories, quantities];
+}
